戻り値型変更 int64_t -> int3_2t Encoder::_getDeltaPulse

+++ encoder.h
@@ -118,7 +118,7 @@ public:
      * @warning 返り値が 64bit値 であることに注意してください．
      * @return Total number of pulses counted so far.
      */
-    inline int64_t getTotalPulse() const { return _total_pulse; }
+    inline int32_t getTotalPulse() const { return _total_pulse; }
 
 
     /**
